+++ open_manipulator_description/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package open_manipulator_description
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+4.1.4 (2026-04-27)
+------------------
+* None
+
 4.1.3 (2026-03-18)
 ------------------
 * Add realsense d405 urdf for OMY-F3M